+++ modules/Material/Icon.qml
@@ -30,9 +30,21 @@ Item {
 
     /*!
        The name of the icon to display.
-       \qmlproperty string name
+       
+       \sa source
     */
     property string name
+
+    /*!
+       A URL pointing to an image to display as the icon. By default, this is
+       a special URL representing the icon named by \l name from the Material Design
+       icon collection or FontAwesome. The icon color will not be changed unless you
+       set the \l colorize property to true or put ".tint." in the filename, in which case
+       the \l color property will be used to color the icon.
+
+       \sa name
+      */
+    property string source: "icon://" + name
 
     property bool valid: name.indexOf("awesome") == 0 
             ? awesomeIcon.valid : image.status == Image.Ready
@@ -42,17 +54,25 @@ Item {
     width: size
     height: size
 
+    property bool colorize: icon.source.indexOf("icon://") == 0 || icon.source.indexOf(".tint.") != -1
+
     Image {
         id: image
 
         anchors.fill: parent
-        visible: false
+        visible: icon.source != "" && !colorize
 
         source: {
-            var list = name.split("/");
-            if (name == "" || list[0] == "awesome")
-                return "";
-            return Qt.resolvedUrl("icons/%1/%2.svg".arg(list[0]).arg(list[1]));
+            if (icon.source.indexOf("icon://") == 0) {
+                var name = icon.source.substring(7)
+                var list = name.split("/");
+
+                if (name == "" || list[0] == "awesome")
+                    return "";
+                return Qt.resolvedUrl("icons/%1/%2.svg".arg(list[0]).arg(list[1]));
+            } else {
+                return icon.source
+            }
         }
 
         sourceSize {
@@ -68,7 +88,7 @@ Item {
         source: image
         color: Theme.alpha(icon.color, 1)
         cached: true
-        visible: image.source != ""
+        visible: icon.source != "" && colorize && image.source != ""
         opacity: icon.color.a
     }  
 
@@ -77,15 +97,20 @@ Item {
 
         anchors.centerIn: parent
         size: icon.size * 0.9
-        visible: icon.name.indexOf("awesome/") == 0
+        visible: name != ""
         color: icon.color
 
         name: {
-            var list = icon.name.split("/")
-            if (list[0] == "awesome") {
-                return list[1]
+            if (icon.source.indexOf("icon://") == 0) {
+                var name = icon.source.substring(7)
+                var list = name.split("/")
+
+                if (list[0] == "awesome") {
+                    return list[1]
+                }
             }
-            return ''
+
+            return ""
         }
     }
 }